Better tutorial (#251)
Addresses #227, #180, #129

I fixed, updated and extended the tutorial. Some stuff in the basic tutorial I moved into an "advanced tutorial". `minimize` is now described in the API page.


I found minor issues in the frontends while doing so, which affect printing information in the console and the notebook. These I fixed as well. I also added a few new tests for the describe function, because I noticed an inconsistency for functors with variable arguments. This is also fixed. Finally, I pep8'd some code using flake8.