Js removal (#320)
Removes JavaScript code in HTML output, to fix rendering bugs of Jupyter Notebooks on Github. This comes at the cost of losing a minor feature - one that not many users will know about - the ability to get a latex version of a few output tables by clicking on a + symbol.