Support annotated model parameters (#839)
This enables support in iminuit and the builtin cost function for
parameter annotations that declare parameter limits. For parameter
annotations to work in Python-3.8, you should use `from __future__
import annotations`.

Parameter limits are declared with the annotated type `Annotated[float, min:max]`
where `min` and `max` are floats including positive or negative infinity or None. 
Annotations of this kind are recognized and automatically used to set parameter limits in `Minuit`.