Clarify meaning of 2D contours in Minuit.draw_mnmatrix (#927)
The method `Minuit.draw_mnmatrix` shows 2D contours that do not align
with the 1D intervals, which is ok. The keyword `cl` sets the confidence
level for the 1D intervals and 2D contours, according to Wilks' theorem,
the projections of the 2D contours are then wider than the 1D intervals.
The docstrings make this point more clear.