Support NumPy 2 (#977)
A new release of iMinuit needs to be made with pybind11 2.12 in order to
support numpy 2. All binaries must be rebuilt with pybind11 2.12+ in
order to support NumPy 2.

That should happen automatically, but I took the opportunity to update a
few things while I was adding a NumPy 2 test. (which I actually forgot
to do, will make draft until I add that!)

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>