Add name argument to all cost functions (#1017)
Closes #941 

So far the iminuit.cost functions only support models that pass each
parameter as a positional argument. This makes it difficult to use
models whose number of arguments should be easy to vary.

To make working with the latter easier, the cost functions now support
models with a variable number of arguments in the form `def model(x,
*par)`. The actual number of arguments and their names are then passed
with the new keyword `name`, which can also be used to override the
model parameter names found by inspecting the model signature.