support sum(w) < 0 in binned fits (#1022)
This patch makes it possible to fit histograms with negative bin entries
with `ExtendedBinnedNLL` and `BinnedNLL`. These occur, for example, when
sweighted samples are histogrammed. This patch implements an extension
of the Bohm-Zech approach for this case. The ideas behind this extension
are compiled in a write-up that can be found in the documentation under
"Studies: Fitting weighted histograms".

The PR also fixes a mistake in `BinnedNLL`, which used the wrong cost
function for weighted histograms, for ordinary histograms there is no
change. This is a side result from thinking deeply about fitting
weighted histograms. The old cost function lead to biased estimates,
while the new one is now almost unbiased and performs even better than
`ExtendedBinnedNLL` in application to weighted histograms.

Other changes:
- Automatic notebook stripping is implemented via pre-commit; so
notebooks that run for a long time are not stripped
- Switch to SVG plots in notebooks

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>