remove packaging dependency (#1027)
We promise to only depend on numpy, so let's remove the packaging
dependency, which is used only for parsing version strings. Correct
parsing of version strings is surprisingly complicated, and I could not
find a standard lib function to do this. Fortunately, we do not need to
support [the full
spec](https://packaging.python.org/en/latest/specifications/version-specifiers/).