Fix ROOT version in docs (#1043)
We opt for a manual update strategy for the ROOT version, because it
changes slowly and an efficient way to do it automatically in CI is
elusive. Instructions were added to the release.md instructions and
conf.py in the doc folder.

#### What is the problem with doing it in CI?

Since the ROOT repo is huge, we want to make shallow clones for git
submodules, but then we cannot determine the relation of the commit to
the closest tag, because that history which connects tag and commit is
not there. Checking out all tags with `git pull --tags` takes a lot of
bandwidth, and is not enough. We also need to check out the history
between the latest tag and the commit. We don't know which tag the
commit is closest to, so the best way would be to iteratively load the
most recent history until the commit is connected to a tag.