Fix Qt widget sizing issues and simplify code (#1075)
Closes #1074 

The layout was changed so the parameters always take the minimal space.
Font size was reduced to 11pt. The size of the result area was increased
so that the full result can be seen on Windows.

The layout was not made user-adjustable, because parameters, buttons,
and the result have fixed sizes anyway, so the only area with an
adjustable size is the plot.

The code was simplified in a couple of places and some blocks refactored
to functions to improve clarity.

Minimum and maximum values for each slider are taken from the
Minuit.limits.