Fixing test suite after regression of scipy.stats.multivariate_normal in scipy-1.16.0 (#1102)
Closes #902

We used to use scipy.stats.multivariate_normal to test fitting in 2D.
For binned fits, this requires the cdf of the model, and there is no
closed form. In the past, that was okay, because the numerical
integration routine in scipy was apparently good enough. But that
changed in 1.16.0 when they switched to a worse-performing
implementation - in regard to numerical accuracy and speed, see
https://github.com/scipy/scipy/issues/23469.

The workaround is to replace scipy.stats.multivariate_normal with a
factorized version which can be implemented based on the pdfs and cdfs
of two 1D normal distributions.

As a drive-by fix, I should have finally fixed
test_Template_with_model_2D. I don't know why, but I used an extreme
coarse binning, which must have caused the numerical issues on some
platforms.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>